backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7
python-dotenv==1.0.0
Flask-Mail==0.9.1
APScheduler==3.10.4
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0

